third_party/blink/web_tests/platform/generic/fast/events/dispatch-synthetic-keyboardevent-no-action-expected.txt

Tests to ensure that default action does not occur for untrusted events.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

 
PASS document.getElementById('radio1').checked is true
PASS document.getElementById('radio2').checked is false
PASS successfullyParsed is true

TEST COMPLETE
